% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{tokenize_simple}
\alias{tokenize_simple}
\title{Tokenise text into a sequence of words}
\usage{
tokenize_simple(x, split = "[[:space:]]+")
}
\arguments{
\item{x}{a character string of length 1}

\item{split}{passed on to \code{\link{strsplit}}}
}
\value{
a character vector with the sequence of words in \code{x}
}
\description{
Tokenise text into a sequence of words. The function uses \code{\link{strsplit}} to split text into words
by using the [:space:] character classes.
}
\examples{
tokenize_simple("This just splits. Text.alongside\nspaces right?")
tokenize_simple("Also .. multiple punctuations or ??marks")
tokenize_simple("Joske  Vermeulen")
}
\seealso{
\code{\link{strsplit}}
}
